move model to sandbox to address CSP issues w ONNX
