+++ examples/extension/public/offscreen.html
@@ -1,2 +1,12 @@
 <!DOCTYPE html>
-<script src="offscreen.js" type="module"></script>
+<html lang="en">
+
+<head>
+    <script src="offscreen.js" type="module"></script>
+</head>
+
+<body>
+    <iframe src="sandbox.html" id="sandbox"></iframe>
+</body>
+
+</html>
