+++ examples/extension/public/sandbox.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<script src="sandbox.js" type="module"></script>
