+++ examples/extension/src/sandbox.js
@@ -1,0 +1,44 @@
+import { pipeline, env } from '@xenova/transformers';
+
+// Skip initial check for local models, since we are not loading any local models.
+env.allowLocalModels = false;
+env.useBrowserCache = false; // for now: caches in sandox might be tricky.
+
+class PipelineSingleton {
+    static task = 'text-classification';
+    static model = 'Xenova/distilbert-base-uncased-finetuned-sst-2-english';
+    static instance = null;
+
+    static async getInstance(progress_callback = null) {
+        if (this.instance === null) {
+            this.instance = pipeline(this.task, this.model, { progress_callback });
+        }
+
+        return this.instance;
+    }
+}
+
+// Create generic classify function, which will be reused for the different types of events.
+const classify = async (text) => {
+    // Get the pipeline instance. This will load and build the model when run for the first time.
+    let model = await PipelineSingleton.getInstance((data) => {
+        // You can track the progress of the pipeline creation here.
+        // e.g., you can send `data` back to the UI to indicate a progress bar
+        // console.log('progress', data)
+    });
+
+    // Actually run the model on the input text
+    let result = await model(text);
+    return result;
+};
+
+window.addEventListener('message', function (event) {
+    console.log('message received in sandbox', event)
+    try {
+        classify(event.data).then(result => {
+            event.ports[0].postMessage({ result });
+        })
+    } catch (e) {
+        event.ports[0].postMessage({ error: e });
+    }
+}, false)
